% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkowitzR.r
\name{MarkowitzR-NEWS}
\alias{MarkowitzR-NEWS}
\title{News for package 'MarkowitzR':}
\description{
News for package \sQuote{MarkowitzR}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\MarkowitzR}{\CRANpkg{MarkowitzR}}
}
\section{Changes in \MarkowitzR{} Version 1.0.1 (2018-05-25) }{

\itemize{
\item move figures around for README on CRAN.
}
}

\section{Changes in \MarkowitzR{} Version 0.9900 (2016-09-15) }{

\itemize{
\item yet again, conform to CRAN rules.
}
}

\section{Changes in \MarkowitzR{} Version 0.1502 (2015-01-26) }{

\itemize{
\item conform to CRAN rules.
}
}

\section{Changes in \MarkowitzR{} Version 0.1403 (2014-06-01) }{

\itemize{
\item fix bug preventing multi-row hedging or constraint matrices.
}
}

\section{\MarkowitzR{} Initial Version 0.1402 (2014-02-14) }{

\itemize{
\item first CRAN release.
}
}

