\name{MarkowitzR-NEWS}
\alias{MarkowitzR-NEWS}
\title{News for package 'MarkowitzR':

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\MarkowitzR}{\CRANpkg{MarkowitzR}}}
\description{
News for package 'MarkowitzR':

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\MarkowitzR}{\CRANpkg{MarkowitzR}}
}
\section{\MarkowitzR{} Initial Version 0.1402 (2014-02-14) }{
  \itemize{ \item first CRAN release. }
}

